search results

+++ src/domain/composer_search_result.rs
@@ -1,8 +1,13 @@
+use serde::{Deserialize, Serialize};
+use sqlx::FromRow;
+
 /// Search result for a composer.
+#[derive(Debug, FromRow, Serialize, Deserialize, Clone)]
+#[serde(rename_all(serialize = "camelCase"))]
 pub struct ComposerSearchResult {
     pub id: i32,
     pub first_name: String,
     pub last_name: String,
     pub slug: String,
-    pub rating: f64,
+    pub rating: f32,
 }
